requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
beanie==1.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic-settings==2.6.0
fastapi-mail==1.4.1
cloudinary==1.41.0
apscheduler==3.10.4
pandas==2.2.3
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langgraph==0.2.39
langchain-chroma==0.1.4
chromadb==0.5.0
python-dotenv==1.0.1
httpx==0.27.0
email-validator==2.2.0
